Add deep reporting base
