+++ default/python/AI/analytics/orders.py
@@ -1,96 +1,64 @@
 from collections import Counter
 from functools import wraps
-from logging import error, info
+from logging import info
 from operator import attrgetter
-from typing import Callable, Dict, Iterator, List, NamedTuple, Tuple
+from typing import Callable, Iterator, List, Tuple
 
 import freeOrionAIInterface as fo
+from analytics.jar import Jar
 from common.print_utils import Number, Table, Text
 
 
-class CallInfo(NamedTuple):
-    args: tuple
-    result: bool
-
-
-class Jar:
-    def __init__(self, name: str):
-        self.name = name
-        self._jar: Dict[int, List[CallInfo]] = {}
-
-    def set_order(self, args: tuple, result: int):
-        if result == 0:
-            result = False
-        elif result == 1:
-            result = True
-        else:
-            error("Unexpected result from issuing order, expect {0, 1} got " % result, stack_info=True)
-            result = True
-
-        self._jar.setdefault(fo.currentTurn(), []).append(CallInfo(args, result))
-
-    def get_turn(self, turn: int) -> List[CallInfo]:
-        return self._jar.get(turn, [])
-
-    def get_total(self):
-        return sum(len(item) for item in self._jar.values())
-
-    def __repr__(self):
-        return "Jar({})".format(self.name)
-
-
 def wrap_order(fun: Callable, call_jar: Jar):
-    @wraps(fun)
-    def wrapper(*args):
-        result = fun(*args)
-        call_jar.set_order(args, result)
-        return result
+  @wraps(fun)
+  def wrapper(*args):
+    result = fun(*args)
+    call_jar.set_order(args, result)
+    return result
 
-    return wrapper
+  return wrapper
 
 
 class CommandTracker:
-    def __init__(self, jar: Jar, fun_name: str):
-        self.fun_name = fun_name
-        self.jar = jar
+  def __init__(self, jar: Jar, fun_name: str):
+    self.fun_name = fun_name
+    self.jar = jar
 
-    def patch(self) -> None:
-        fun = getattr(fo, self.fun_name)
-        wrapped = wrap_order(fun, self.jar)
-        setattr(fo, self.fun_name, wrapped)
+  def patch(self) -> None:
+    fun = getattr(fo, self.fun_name)
+    wrapped = wrap_order(fun, self.jar)
+    setattr(fo, self.fun_name, wrapped)
 
 
 def get_issuers() -> Iterator[Tuple[str, str]]:
-    for attr in dir(fo):
-        if attr.startswith("issue") and attr.endswith("Order"):
-            yield attr[5:-5], attr
+  for attr in dir(fo):
+    if attr.startswith("issue") and attr.endswith("Order"):
+      yield attr[5:-5], attr
 
 
 class Tracker:
-    def __init__(self):
+  def __init__(self):
+    self.jars: List[Jar] = []
+    for name, issuer in get_issuers():
+      jar = Jar(name)
+      self.jars.append(jar)
+      tracker = CommandTracker(jar, issuer)
+      tracker.patch()
+    self.jars.sort(key=attrgetter("name"))
 
-        self.jars: List[Jar] = []
+  def report_turn(self):
+    turn = fo.currentTurn()
+    table = Table(
+      Text('Order'),
+      Number("Issued", precession=0, placeholder=" ", total=True),
+      Number("Failed", precession=0, placeholder=" ", total=True),
+      Number("Total", precession=0, placeholder=" ", total=True),
+      table_name="Issuing orders analytics for turn {}".format(turn),
+    )
 
-        for name, attr in get_issuers():
-            jar = Jar(name)
-            self.jars.append(jar)
-            tracker = CommandTracker(jar, attr)
-            tracker.patch()
-        self.jars.sort(key=attrgetter("name"))
-
-    def report_turn(self):
-        turn = fo.currentTurn()
-        table = Table(
-            Text('Order'),
-            Number("Issued", precession=0, placeholder=" ", total=True),
-            Number("Failed", precession=0, placeholder=" ", total=True),
-            Number("Total", precession=0, placeholder=" ", total=True),
-            table_name="Issuing orders analytics for turn {}".format(turn),
-        )
-
-        for jar in self.jars:
-            results = Counter(item.result for item in jar.get_turn(turn))
-            success = results[True]
-            fail = results[False]
-            table.add_row(jar.name, success, fail, jar.get_total())
-        table.print_table(info)
+    for jar in self.jars:
+      results = Counter(item.result for item in jar.get_turn(turn))
+      success = results[True]
+      fail = results[False]
+      table.add_row(jar.name, success, fail, jar.get_total())
+    table.print_table(info)
