+++ default/python/AI/analytics/report/base_handler.py
@@ -1,0 +1,6 @@
+from analytics.jar import CallInfo
+
+
+def base_reported(call_info: CallInfo) -> str:
+  res = "OK" if call_info.result else "Fail"
+  return f'[{call_info.name}]: {call_info.args}: {res}'
